requirements.txt
anthropic>=0.25.0
openai>=1.14.0
streamlit>=1.32.0
python-dotenv>=1.0.0
